pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "confluence-exporter"
version = "0.1.0"
description = "Export a Confluence space to PDF/DOCX/Markdown/HTML with attachments, merging, and an interactive UI."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["confluence", "atlassian", "exporter", "pdf", "documentation", "notebooklm"]
authors = [
  { name = "Leano Chiodo" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Documentation",
  "Topic :: Utilities",
]

# ------- Runtime dependencies (always needed) -------
dependencies = [
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=4.9",
  "markdownify>=0.11",
  "python-docx>=1.1",
  "Pillow>=10.0",
  "pypdf>=4.0",
  "xhtml2pdf>=0.2.11",     # pure-Python fallback engine
  "rich>=13.7",            # pretty UI / progress bars / prompts
  "typer>=0.12",           # CLI framework
]

[project.optional-dependencies]
playwright = ["playwright>=1.40"]
weasyprint = ["weasyprint>=60.0"]
all = [
  "playwright>=1.40",
  "weasyprint>=60.0",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "build>=1.2",
  "twine>=5.0",
]

[project.urls]
Homepage = "https://github.com/LeoChi/confluence-exporter"
Repository = "https://github.com/LeoChi/confluence-exporter"
Issues = "https://github.com/LeoChi/confluence-exporter/issues"
Changelog = "https://github.com/LeoChi/confluence-exporter/blob/main/CHANGELOG.md"

# ------- Entry points -------
[project.scripts]
confluence-exporter = "confluence_exporter.cli:app"
cfx = "confluence_exporter.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_exporter"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/examples",
  "/README.md",
  "/LICENSE",
  "/CHANGELOG.md",
]

# ------- Tools -------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "N", "SIM"]
ignore = [
  "E501",  # line-too-long handled by formatter
  "B008",  # function call in default arg (typer uses this pattern)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
